feat: add ApprovedTravelsPrinter class for generating PDF reports
